backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
openai==1.3.7
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
pydantic-settings==2.1.0
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0